Put integrator into new file
